Separando os arquivos para poder acessar o banco nos demais módulos
