+++ ouvICEx/admin/posts.py
@@ -1,7 +1,8 @@
 from flask import Flask, Blueprint, render_template
+from database.posts import posts
 
-b_posts = Blueprint("b_posts", __name__, static_folder="static", template_folder="templates")
+admin_view = Blueprint("admin_view", __name__, static_folder="static", template_folder="templates")
 
-@b_posts.route("/posts")
-def posts():
-    return render_template("posts.html")
+@admin_view.route("/view")
+def view():
+    return render_template("view.html", posts=posts.query.all())
