+++ ouvICEx/admin/templates/view.html
@@ -2,4 +2,7 @@
 {% block title %} Postagens {% endblock %}
 {% block content %}
     <p> Pagina do administrador para gerenciar as postagens </p>
+    {% for p in posts %}
+        <p> {{p.post}} </p>
+    {% endfor %}
 {% endblock %}
