+++ ouvICEx/init_flask.py
@@ -1,31 +1,12 @@
 from flask import Flask, render_template
-from flask_sqlalchemy import SQLAlchemy
 
-from admin.posts import b_posts
+from database.posts import db
+from admin.posts import admin_view
 
 app = Flask(__name__)
-app.register_blueprint(b_posts, url_prefix="/admin")
+app.register_blueprint(admin_view, url_prefix="/admin")
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///posts.sqlite3'
 app.config['SQLALCHEMY_TRACK_NOTIFICATIONS'] = False
-
-db = SQLAlchemy(app)
-
-class posts(db.Model):
-    _id = db.Column("id", db.Integer, primary_key=True)
-    post = db.Column("post", db.String(500))
-    date = db.Column("date", db.Date)
-    author_dep = db.Column("author_dep", db.String(50))
-    ref_dep = db.Column("ref_dep", db.String(50))
-    context_t = db.Column("context_t", db.String(20))
-    situation_t = db.Column("situation_t", db.String(20))
-
-    def __init__(self, post, date, author_dep, ref_dep, context_t, situation_t):
-        self.post = post
-        self.date = date
-        self.author_dep = author_dep
-        self.ref_dep = ref_dep
-        self.context_t = context_t
-        self.situation_t = situation_t
 
 @app.route("/")
 def home():
@@ -33,5 +14,5 @@ def home():
 
 
 if __name__ == "__main__":
-    db.create_all()
+    db.init_app(app)
     app.run(debug=True)
